add bootstrap to navbar
